# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: EdgeAI Control Plane
  version: 0.1.0
  description: Profile → Device/Node → Workflow DAG → Run/Task/Attempt와 검증된 결과 조회 API입니다.
    실제 Kubernetes 실행은 서버의 runtime 설정이 활성화된 환경에서 동작합니다. 쓰기 요청은 실제 DB에 반영되고 Swagger의 CSRF 토큰/쿠키는 자동 연결됩니다.
    복원 DB의 조회 전용 점검 모드에서는 실행 기능이 비활성화되고 CSRF가 유효한 쓰기 요청도 403으로 거절됩니다.
    일반 관리 API의 변경 요청은 먼저 감사 접수 기록을 저장하며 X-EdgeAI-Audit-Id를 반환합니다. 접수 저장 실패는503/AUDIT_STORE_UNAVAILABLE이며 변경을 실행하지 않습니다.
    감사의 HTTP 결과는 비동기 작업 성공을 뜻하지 않으며, 결과가 없는 접수 기록은 OUTCOME_UNKNOWN입니다. 복구 점검 모드에서는 감사 쓰기도 비활성화됩니다.
servers:
- url: /
security: []
paths:
...
  /api/v1/devices/{deviceId}:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema:
        type: string
        format: uuid
    get:
      tags:
      - Device
      operationId: getDevice
      summary: 장치 상세·연결 이력·최근 관측 조회
      description: 장치와 최근 연결/세션/관측 각각 최대 20개를 반환합니다. 시간은 UTC이며 신선도는 서버 수신 시각으로 판단합니다. raw 스트림은 저장하지 않습니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Device
      operationId: updateDevice
      summary: 장치 표시 이름 수정
      description: 상세 조회에서 받은 revision과 새 displayName을 전송합니다. 동시에 수정되어 revision이 바뀌면 409입니다. Profile 참조·key·sourceMode는 변경하지
        않습니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicePatch'
    delete:
      tags:
      - Device
      operationId: releaseDevice
      summary: 장치 해제 및 세션 종료
      description: 장치를 RELEASED로 표시하고 활성 attachment/session을 닫습니다. 등록된 VD 또는 종료 대기 중인 이전 VD 실행이 원본으로 사용하면 DEVICE_IN_USE(409)입니다. 원본 연결 해제와 이전 실행의 종료 확인이 모두 필요합니다. 과거 이력을 보존하며 반복 해제는 같은 상태를 반환합니다. 해제 후 연결·세션·관측·수정 요청은 409입니다.
      responses:
        '200':
          description: 요청 성공. 재전송은 기존 객체를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: 필드·UUID·페이지·JSON 형식을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 장치·노드·참조 Profile이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: DB 연결 장애. 복구 후 재시도하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '409':
          description: 현재 상태·버전·세션·sequence 충돌. 상세 조회 후 요청을 확인하세요.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: 요청 크기 제한을 초과했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - csrfToken: []
  /api/v1/devices/{deviceId}/registration:
    parameters:
    - in: path
      name: deviceId
      required: true
      schema: {type: string, format: uuid}
    delete:
      operationId: deleteDeviceRegistration
      tags: [Device]
      summary: 사용하지 않는 장치 등록 삭제
      description: 활성 연결·세션이나 VD·실행 이력 참조가 있으면 DEVICE_IN_USE(409)로 거절합니다. 미사용 장치와 그 장치의 닫힌 연결·세션·관측 이력만 삭제하며 Kubernetes 노드는 변경하지 않습니다.
      security:
      - csrfToken: []
      responses:
        '204':
          description: 장치 등록 삭제 완료
        '400':
          $ref: '#/components/responses/Invalid'
        '403':
          description: CSRF 토큰·세션 쿠키 누락/불일치
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: DEVICE_IN_USE — 활성 연결·세션 또는 다른 자원의 참조가 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/v1/virtual-devices:
    post:
      tags: [VD]
      operationId: createVirtualDevice
      summary: VD 식별자와 원본 장치 연결 등록
      description: >-
        edgeai.vd/v1 규격의 VD Profile과 실제 DEVICE/SERVICE 버전을 검증하고 원본 연결을 원자적으로 저장합니다.
        같은 key와 생성 입력은 기존 VD를 반환하고 다른 입력은409입니다. 해제 뒤 재전송도 기존 VD이며 재활성화하지 않습니다.
        현재 REGISTERED는 논리 등록만 완료된 상태입니다. 실행은 별도 provision 요청으로 시작하고 execution에서 준비 상태를 확인합니다. VD 경유 Task 실행은 후속 연결 범위입니다.
      security: [{csrfToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VirtualDeviceCreate'}
      responses:
        '201':
          description: 논리 VD와 원본 연결 생성. 실제 Runtime 준비 완료를 뜻하지 않습니다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualDevice'}
        '200':
          description: 같은 생성 입력의 기존 VD
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualDevice'}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '403': {description: CSRF 쿠키/토큰 필요}
        '404': {$ref: '#/components/responses/VirtualDeviceError'}
        '409': {$ref: '#/components/responses/VirtualDeviceError'}
        '413': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
    get:
      tags: [VD]
      operationId: listVirtualDevices
      summary: 등록·해제된 VD 목록 조회
      description: key 순서로 논리 VD를 조회합니다. REGISTERED를 Ready로 해석하지 마세요. 해제 이력도 보존합니다.
      parameters:
      - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}, description: 한 페이지 수}
      - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: 건너뛸 수}
      responses:
        '200':
          description: VD 목록과 다음 offset
          content:
            application/json:
              schema:
                type: object
                required: [items, nextOffset]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/VirtualDevice'}}
                  nextOffset: {type: [integer, 'null']}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
  /api/v1/virtual-devices/{vdId}:
    parameters:
    - {name: vdId, in: path, required: true, schema: {type: string, format: uuid}, description: 원본·실행 위치 변경에도 유지할 VD UUID}
    get:
      tags: [VD]
      operationId: getVirtualDevice
      summary: VD 설정과 활성·과거 원본 연결 조회
      description: 최신 revision, 불변 Profile 참조와 source 이력을 조회합니다. closedAt이 null인 연결만 현재 원본입니다. 이력은 최신100개이며 더 있으면 sourceHistoryTruncated=true입니다.
      responses:
        '200':
          description: VD 상세
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualDeviceDetail'}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '404': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
    patch:
      tags: [VD]
      operationId: updateVirtualDevice
      summary: revision을 확인하여 VD 이름·원본·배치 의도 수정
      description: >-
        이름·sources 전체 집합·placement를 함께 보내세요. Profile 참조와 vdId는 유지합니다.
        변경하지 않은 slot은 같은 연결 ID를 유지하고 바뀐 원본은 기존 이력을 닫고 새 연결을 만듭니다.
        오래된 revision·해제된 VD는409입니다. 현재 논리 설정 수정이며 실제 Runtime 전환 성공을 뜻하지 않습니다.
        관리 중인 실행이 있으면 원본·배치 변경과 교체 Operation을 같은 트랜잭션에 기록합니다. 이름만 바꾸면 실행을 교체하지 않습니다.
      security: [{csrfToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VirtualDeviceUpdate'}
      responses:
        '200':
          description: 최신 VD 설정. 같은 내용이면 revision도 유지합니다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualDevice'}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '403': {description: CSRF 쿠키/토큰 필요}
        '404': {$ref: '#/components/responses/VirtualDeviceError'}
        '409': {$ref: '#/components/responses/VirtualDeviceError'}
        '413': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
    delete:
      tags: [VD]
      operationId: releaseVirtualDevice
      summary: VD 논리 해제와 원본 연결 종료
      description: VD를 RELEASED로 표시하고 활성 source를 닫습니다. 관리 중인 실행이 있으면 종료 Operation을 함께 기록하며 완료 전까지 이전 원본 Device 해제를 막습니다. 물리 Device는 해제하지 않으며 이력과 vdId를 유지합니다. 반복 해제는 같은 상태를 반환합니다. 응답은 논리 해제이며 실제 Pod 종료 완료를 뜻하지 않습니다. execution의 pendingOperation과 Operation 조회에서 실제 종료를 확인하세요.
      security: [{csrfToken: []}]
      responses:
        '200':
          description: 해제된 VD
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualDevice'}
        '400': {$ref: '#/components/responses/VirtualDeviceError'}
        '403': {description: CSRF 쿠키/토큰 필요}
        '404': {$ref: '#/components/responses/VirtualDeviceError'}
        '503': {$ref: '#/components/responses/VirtualDeviceError'}
...
  /api/v1/virtual-devices/{vdId}/registration:
    delete:
      operationId: deleteVirtualDeviceRegistration
      summary: 해제된 미사용 가상 장치 등록 영구 삭제
      tags: [가상 장치]
      description: 실행·작업 참조가 없는 RELEASED VD와 닫힌 원본 연결 기록만 삭제한다. 원본 물리 장치와 프로필은 보존한다. 실행 이력은 삭제하지 않으며 참조가 남으면409로 거절한다.
      parameters:
      - {name: vdId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '204': {description: 등록 삭제 완료}
        '403': {description: CSRF 검증 실패}
        '404': {description: VD 없음}
        '409': {description: 미해제 또는 실행·작업 참조가 남음}
        '503': {description: 저장소 사용 불가}
components:
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: 쓰기 요청은 세션 쿠키와 함께 이 헤더가 필요합니다. Swagger는 자동 발급·전달하므로 이 입력칸을 비워 두세요. 직접 호출할 때는 getCsrfToken을 먼저 사용합니다.
...
  responses:
...
    VirtualDeviceError:
      description: 잘못된 입력400·없는 VD/Profile/Device/Node404·중복 키/revision/원본 충돌409·64KiB 초과413·DB 장애503입니다. code와 message를 확인하세요.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Invalid:
      description: 종류·키·버전·페이지·JSON 형식이 잘못되었습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_PROFILE
            message: 종류, 키, 버전과 JSON 규격을 확인하세요. 규격은 비어 있지 않은 객체여야 합니다.
...
    NotFound:
      description: 해당 발행 버전이 없습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_NOT_FOUND
            message: 해당 Profile 버전을 찾을 수 없습니다.
    Unavailable:
      description: DB 연결 등 저장소 문제로 요청을 처리할 수 없습니다. 복구 후 재시도하세요.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: PROFILE_STORE_UNAVAILABLE
            message: Profile 저장소에 연결할 수 없습니다. 잠시 후 다시 시도하세요.
  schemas:
...
    ProfileKey:
      type: string
      maxLength: 100
      pattern: ^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$
      description: 소문자 영문으로 시작하는 프로필 키. 영문 소문자·숫자와 구분자 점/밑줄/하이픈을 사용하며 최대 100자입니다.
      example: temperature-sensor
...
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INVALID_PROFILE
          - PROFILE_CONFLICT
          - PROFILE_NOT_FOUND
          - PROFILE_STORE_UNAVAILABLE
          - PAYLOAD_TOO_LARGE
          - INVALID_DEVICE
          - DEVICE_CONFLICT
          - DEVICE_NOT_FOUND
          - NODE_NOT_FOUND
          - NODE_NOT_READY
          - DEVICE_RELEASED
          - STALE_SESSION
          - OBSERVATION_CONFLICT
          - DEVICE_STORE_UNAVAILABLE
          - INVALID_WORKFLOW
          - WORKFLOW_NOT_FOUND
          - WORKFLOW_CONFLICT
          - RUN_NOT_FOUND
          - TASK_NOT_FOUND
          - IDEMPOTENCY_CONFLICT
          - CANNOT_CANCEL
          - STREAM_NOT_IMPLEMENTED
          - STREAM_EXECUTION_POLICY_UNSUPPORTED
          - STREAM_RECOVERY_UNSUPPORTED
          - STREAM_SOURCE_INACTIVE
          - WORKFLOW_STORE_UNAVAILABLE
          - INVALID_TASK_ID
          - RESULT_STORE_UNAVAILABLE
          - RUNTIME_DISABLED
          - REMOTE_DISABLED
          - REMOTE_PROVIDER_NOT_FOUND
          - REMOTE_CONFIGURATION_CHANGED
          - REMOTE_TELEMETRY_UNSUPPORTED
          - VD_EXECUTION_DISABLED
          - VD_NOT_READY
          - VD_SERVICE_MISMATCH
          - VD_AUTOMATIC_OFFLOAD_UNSUPPORTED
          - OFFLOAD_SOURCE_CHANGED
          - OFFLOAD_LIMIT
          - OFFLOAD_RECOVERY_UNSUPPORTED
          - OFFLOAD_TARGET_INVALID
          - OPERATION_NOT_FOUND
          - INVALID_VIRTUAL_DEVICE
          - VD_NOT_FOUND
          - VD_CONFLICT
          - VD_RELEASED
          - VD_SOURCE_INCOMPATIBLE
          - VD_STORE_UNAVAILABLE
          - DEVICE_IN_USE
        message:
          type: string
...
    DevicePatch:
      type: object
      additionalProperties: false
      required:
      - revision
      - displayName
      properties:
        revision:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        displayName:
          type: string
          minLength: 1
          maxLength: 128
...
    Device:
      type: object
      additionalProperties: false
      required:
      - id
      - key
      - displayName
      - profileVersionId
      - sourceMode
      - state
      - revision
      - sessionEpoch
      - connectionStatus
      - lastSeenAt
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        key:
          $ref: '#/components/schemas/ProfileKey'
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        profileVersionId:
          type: string
          format: uuid
        sourceMode:
          type: string
          enum:
          - LIVE
          - REPLAY
          - SYNTHETIC
        state:
          type: string
          enum:
          - ACTIVE
          - RELEASED
        revision:
          type: integer
          minimum: 0
        sessionEpoch:
          type: integer
          minimum: 0
        connectionStatus:
          type: string
          enum:
          - UNKNOWN
          - ONLINE
          - OFFLINE
          - STALE
          - RELEASED
        lastSeenAt:
          type:
          - string
          - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceAttachment:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - nodeId
      - port
      - attachedAt
      - detachedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        nodeId:
          type: string
          format: uuid
        port:
          type: string
        attachedAt:
          type: string
          format: date-time
        detachedAt:
          type:
          - string
          - 'null'
          format: date-time
    DeviceSession:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - bootId
      - epoch
      - lastSequence
      - openedAt
      - closedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        bootId:
          type: string
          format: uuid
        epoch:
          type: integer
          minimum: 1
        lastSequence:
          type: integer
          minimum: -1
        openedAt:
          type: string
          format: date-time
        closedAt:
          type:
          - string
          - 'null'
          format: date-time
    DeviceObservation:
      type: object
      additionalProperties: false
      required:
      - id
      - deviceId
      - sessionId
      - sequence
      - observedAt
      - status
      - attributes
      - receivedAt
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        observedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - ONLINE
          - OFFLINE
        attributes:
          type: object
          additionalProperties: true
        receivedAt:
          type: string
          format: date-time
    DeviceDetail:
      type: object
      additionalProperties: false
      required:
      - device
      - attachments
      - sessions
      - observations
      properties:
        device:
          $ref: '#/components/schemas/Device'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/DeviceAttachment'
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/DeviceSession'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/DeviceObservation'
...
    VirtualDeviceSources:
      type: array
      maxItems: 16
      description: Profile에 선언한 sourceKey마다 Device 하나. 필수 원본은 모두 지정합니다.
      items:
        type: object
        additionalProperties: false
        required: [sourceKey, deviceId]
        properties:
          sourceKey: {$ref: '#/components/schemas/ProfileKey'}
          deviceId: {type: string, format: uuid}
    VirtualDevicePlacement:
      oneOf:
      - type: object
        additionalProperties: false
        required: [mode]
        properties:
          mode: {const: AUTO}
      - type: object
        additionalProperties: false
        required: [mode, nodeId]
        properties:
          mode: {const: NODE}
          nodeId: {type: string, format: uuid}
      description: 후속 VD provision에 사용할 배치 의도. Node는 존재해야 하며 실제 배치는 scheduler가 담당합니다.
    VirtualDeviceCreate:
      type: object
      additionalProperties: false
      required: [key, displayName, profileVersionId, sources, placement]
      properties:
        key: {$ref: '#/components/schemas/ProfileKey'}
        displayName: {type: string, minLength: 1, maxLength: 128}
        profileVersionId: {type: string, format: uuid, description: edgeai.vd/v1 규격의 발행된 VD Profile UUID}
        sources: {$ref: '#/components/schemas/VirtualDeviceSources'}
        placement: {$ref: '#/components/schemas/VirtualDevicePlacement'}
    VirtualDeviceUpdate:
      type: object
      additionalProperties: false
      required: [revision, displayName, sources, placement]
      properties:
        revision: {type: integer, format: int64, minimum: 0, maximum: 9007199254740991}
        displayName: {type: string, minLength: 1, maxLength: 128}
        sources: {$ref: '#/components/schemas/VirtualDeviceSources'}
        placement: {$ref: '#/components/schemas/VirtualDevicePlacement'}
    VirtualDevice:
      type: object
      required: [id, key, displayName, profileVersionId, serviceProfileVersionId, state, revision, placement, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        key: {$ref: '#/components/schemas/ProfileKey'}
        displayName: {type: string}
        profileVersionId: {type: string, format: uuid}
        serviceProfileVersionId: {type: string, format: uuid}
        state: {type: string, enum: [REGISTERED, RELEASED], description: 현재 논리 등록/해제. 실제 Runtime Ready 상태가 아닙니다.}
        revision: {type: integer, format: int64, minimum: 0, maximum: 9007199254740991}
        placement: {$ref: '#/components/schemas/VirtualDevicePlacement'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    VDSourceBinding:
      type: object
      required: [id, vdId, sourceKey, deviceId, deviceProfileVersionId, sourceMode, openedRevision, closedRevision, openedAt, closedAt]
      properties:
        id: {type: string, format: uuid}
        vdId: {type: string, format: uuid}
        sourceKey: {type: string}
        deviceId: {type: string, format: uuid}
        deviceProfileVersionId: {type: string, format: uuid}
        sourceMode: {type: string, enum: [LIVE, REPLAY, SYNTHETIC]}
        openedRevision: {type: integer, format: int64}
        closedRevision: {type: [integer, 'null'], format: int64}
        openedAt: {type: string, format: date-time}
        closedAt: {type: [string, 'null'], format: date-time}
    VirtualDeviceDetail:
      type: object
      required: [vd, activeSources, sourceHistory, sourceHistoryTruncated]
      properties:
        vd: {$ref: '#/components/schemas/VirtualDevice'}
        activeSources: {type: array, maxItems: 16, items: {$ref: '#/components/schemas/VDSourceBinding'}}
        sourceHistory: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/VDSourceBinding'}}
        sourceHistoryTruncated: {type: boolean}
...